refactor: extract edit planner and atomic write to a sibling module
Move _EditPlan, _plan_edits_sync, _write_sync, and the edit-error
constants into tools/file_system/_edit_plan.py so edit_file.py stays
within its module-size budget after the output-policy delta guard grew.
The planner and atomic write are pure functions with no dependency on the
tool class, so the split is a clean seam rather than a size-driven trim.
